Create blog app and moved code in it
