+++ blog/templates/base2.html
@@ -14,7 +14,7 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
         crossorigin="anonymous">
-    <link rel="stylesheet" href="{% static 'css/booking_list_view.css' %}">
+    <link rel="stylesheet" href="{% static 'css/style.css' %}">
 </head>
 
 <body class="bg-light">
@@ -52,42 +52,8 @@
             <img src="https://cdn.pixabay.com/photo/2016/12/07/23/41/norway-1890655_960_720.jpg" width="100%"
                 height="auto" alt="Green valley with houses" class='container' />
         </section> -->
-        <section id="home" Class="img-fluid">
-            <div class="headings">
-                <h2>Your Bookings</h2>
-                <h3>Welcome dear {{ user }}</h3>
-            </div>
-        </section>
-        <div class="container mt-5 pt-5 card-group bookings">
-            <div class="row">
-                {% for booking in booking_list %}
-                <div class="col-md-3 card booking">
-                    <img src="https://res.cloudinary.com/attstar/image/upload/v1677787069/media/hotels/hotel_room_4_oqrnai.jpg" class="card-img-top" alt="hotel room">
-                    <div class="card-body">
-                        <h5 class="card-title"></h5>
-                        <div class="category col-md-5">
-                            <h6>{{ booking.get_room_category }} Room </h6>
-                        </div>
-                        <div class="from">
-                            <h6>From: {{ booking.check_in }}</h6>
-                        </div>
-                        <div class="to">
-                            <h6>To: {{ booking.check_out }}</h6>
-                        </div>
-                        <div class="beds">
-                            <h6>Beds: {{ booking.room.beds }}</h6>
-                        </div>
-                        <div class="capacity">
-                            <h6>Capacity: {{ booking.room.capacity }}</h6>
-                        </div>
-                    </div>
-                    <div class="cancel">
-                        <a href="{{ booking.get_cancel_booking_url }}"><i class="far fa-2x fa-times-circle"></i><h6>Cancel Booking</h6></a>
-                    </div>
-                </div>
-                {% endfor %}
-            </div>
-        </div>    
+        {% block content %}
+        {% endblock %}
     </article>
     <footer class="mt-5">
         <div class="link d-flex justify-content-center">
