+++ blog/templates/post_detail.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %} {% block content %}
+{% extends 'base2.html' %} {% block content %}
 {% load crispy_forms_tags %}
 
 <div class="masthead">
