+++ blog/urls.py
@@ -1,0 +1,8 @@
+from . import views
+from django.urls import path
+
+urlpatterns = [
+    path('blog', views.PostList.as_view(), name='blog'),
+    path('blog/<slug:slug>/', views.PostDetail.as_view(), name='post_detail'),
+    path('like/<slug:slug>', views.PostLike.as_view(), name='post_like'),
+]
