+++ home/admin.py
@@ -1,28 +1,8 @@
 from django.contrib import admin
-from .models import RoomCategory, Room, Booking, Post, Comment
-from django_summernote.admin import SummernoteModelAdmin
+from .models import RoomCategory, Room, Booking
+#from django_summernote.admin import SummernoteModelAdmin
 
 # Register your models here.
-
-@admin.register(Post)
-class PostAdmin(SummernoteModelAdmin):
-
-    list_display = ('title', 'slug', 'status', 'created_on')
-    search_fields = ['title', 'content']
-    list_filter = ('status', 'created_on')
-    prepopulated_fields = {'slug': ('title',)}
-    summernote_fields = ('content',)
-
-
-@admin.register(Comment)
-class CommentAdmin(admin.ModelAdmin):
-    list_display = ('name', 'body', 'post', 'created_on', 'approved')
-    list_filter = ('approved', 'created_on')
-    search_fields = ('name', 'email', 'body')
-    actions = ['approve_comments']
-
-    def approve_comments(self, request, queryset):
-        queryset.update(approved=True)
 
 admin.site.register(RoomCategory)
 admin.site.register(Room)
