+++ home/static/css/booking_list_view.css
@@ -1,23 +1,31 @@
-/* *{
+*{
     padding:0;
     margin:0;
     box-sizing: border-box;
 }
 body{
+    min-height: 100vh;
     font-family: sans-serif;
     line-height: 1.5;
     font-size: 1.2rem;
     text-align: center;
-    flex: 1;
     display: flex;
+    flex-direction: column;
     flex-flow: row-reverse wrap;
     justify-content: space-evenly;
     align-items: center;
-    /* background-image: url('https://cdn.pixabay.com/photo/2016/12/07/23/41/norway-1890655_960_720.jpg'); */
-/* }
+    background-image: url('https://res.cloudinary.com/attstar/image/upload/v1701656637/s2os1iin1qhziamwzeqh.png');
+    background-repeat: no-repeat;
+    background-position: top center;
+    
+}
 
 h1 {
   font-family: 'Caveat', cursive;
+}
+
+h2, h3 {
+    color: #f7f2f2;
 }
 .navbar{
     list-style: none;
@@ -86,4 +94,3 @@ h1,h2{
     color: rgb(206, 43, 43);
     
 }
- */ */
