+++ home/templates/home.html
@@ -23,7 +23,7 @@
     </div>
 </section>
 
-<section class="search-container mb-5 pb-5">
+<!-- <section class="search-container mb-5 pb-5">
     <div class="container .bg-secondary">
         <div class="col-lg-9 px-4 mb-5">
             <form>
@@ -53,7 +53,7 @@
                 </div>
             </form>
 
-            <!-- Hotel Detail Column -->
+            Hotel Detail Column 
             <div class="col-12 mt-3 left">
                 <div class="row search-card-result .bg-dark">
                     {% for hotels_obj in hotels_objs %}
@@ -62,7 +62,7 @@
                         <div class="card mb-4">
                             <div class="card-body">
                                 <div class="image-container">
-                                    <!-- put images setting in hotel -->
+                                     put images setting in hotel 
                                     {% if "placeholder" in hotel.featured_image.url %}
                                     <img class="card-img-top"
                                         src="https://res.cloudinary.com/attstar/image/upload/v1670873330/2_tur7do.png">
@@ -109,7 +109,7 @@
         {% endif %}
     </div>
     </div>
-</section>
+</section> -->
 
 <script>
