+++ home/urls.py
@@ -1,14 +1,15 @@
-from django.urls import path
+from django.urls import path, include
 from . import views
-from .views import BookingListView, BookingFormView, CancelBookingView
+# from .views import BookingListView, BookingFormView, CancelBookingView
 from django.conf import settings
 
 urlpatterns = [
-    path('', BookingFormView.as_view(), name='home'),
-    path('booking_list/', BookingListView.as_view(), name='BookingListView'),
-    path('booking/cancel/<pk>', CancelBookingView.as_view(),
+    path('', views.BookingFormView.as_view(), name='home'),
+    path('booking_list/', views.BookingListView.as_view(), name='BookingListView'),
+    path('booking/cancel/<pk>', views.CancelBookingView.as_view(),
          name='CancelBookingView'),
-    path('blog/', views.PostList.as_view(), name='blog'),
-    path('<slug:slug>/', views.PostDetail.as_view(), name='post_detail'),
-    path('like/<slug:slug>', views.PostLike.as_view(), name='post_like'),
+    path('', include('blog.urls')),
+#     path('blog/', views.PostList.as_view(), name='blog'),
+#     path('<slug:slug>/', views.PostDetail.as_view(), name='post_detail'),
+#     path('like/<slug:slug>', views.PostLike.as_view(), name='post_like'),
 ]
