+++ requirements.txt
@@ -1,12 +1,14 @@
 asgiref==3.7.2
+bootstrap4==0.1.0
 cloudinary==1.34.0
+crispy-bootstrap4==2023.1
 dj-database-url==2.1.0
 dj-rest-auth==4.0.1
-Django==3.2.20
+Django==5.0
 django-allauth==0.54.0
 django-cloudinary-storage==0.3.0
 django-cors-headers==4.2.0
-django-crispy-forms==2.0
+django-crispy-forms==1.14.0
 django-filter==23.2
 django-summernote==0.8.20.0
 djangorestframework==3.14.0
@@ -20,3 +22,4 @@ python3-openid==3.2.0
 pytz==2023.3
 requests-oauthlib==1.3.1
 sqlparse==0.4.4
+urllib3==1.26.18
